add resolveInterface prop
functions similarly to the other `resolve*` props.
we provide a `PropsStore` class that can intelligently resolve props
including inheritance.

this is necessary because the previous impl used a relative path
that does not exist in internal repo.
